// src/types/customer/contact.ts
import type { BASE_OBJECT_RESPONSE } from '@/types'

export type CONTACT_RESPONSE = Nullable<{
  id: number
  address_1: string
  address_2: string
  created_at: string
  description: string
  email: string
  first_name: string
  last_name: string
  name: string
  note: string
  phone: string
  status: number
  updated_at: string
}>